requirements.txt
streamlit
pillow
rembg[cli]
onnxruntime
requests
watchdog